--- cfno/configs/fno3d.yaml
@@ -1,61 +1,0 @@
-default: &DEFAULT
-run: 1
-dim: 'FNO3D'
-exit_signal_handler: True
-exit_duration_in_mins: None
-resume_from_checkpoint: None
-save_path: 'results/strategy4'
-verbose: True
-
-data:
-  train_data_path: 'RayleighBernardConvection/processed_data/RBC2D_NX4_64_NZ1_64_TI50_TF100_Pr1_Ra1e7_dt0_1.h5'
-  val_data_path: 'RayleighBernardConvection/processed_data/RBC2D_NX4_64_NZ1_64_TI50_TF100_Pr1_Ra1e7_dt0_1.h5'
-  train_samples: 40
-  val_samples: 20
-  dt: 0.1
-  nx: 256
-  ny: 64
-  xStep: 1
-  yStep: 1
-  tStep: 1
-  start_index: 500
-  stop_index: 1000
-  timestep: 1
-  batch_size: 40
-  rayleigh_number: 1e7
-  prandtl_number: 1
-  subdomain: True
-  subdomain_args: 
-    start_time: 50.0
-    stop_time: 100.0
-    ndomain_x: 4
-    ndomain_y: 1
-
-FNO:
-  modes: 12
-  width: 20
-  T_in: 1
-  T: 1
-  
-opt:
-  epochs: 3
-  learning_rate: 0.00039
-  weight_decay: 1e-5
-  scheduler_step: 100.0
-  scheduler_gamma: 0.98
-  T_max: 3 # For cosine only, typically take epochs
-  
-inference:
-  do_inference: True
-  model_checkpoint: 'results/strategy3/rbc_fno3d_N10_epoch3_m12_w20_bs1_dt0.1_tin1_run1/checkpoint/model_checkpoint_2.pt'
-  output_timesteps: 1
-  output_error: True
-  save_inference: True
-  plot_cross_section: True
-  test_data_path: RayleighBernardConvection/processed_data/RBC2D_NX256_NZ64_TI0_TF150_Pr1_Ra1e7_dt0_1.h5
-  test_start_index: 50
-  test_stop_index: 1
-  test_timestep: 1
-  test_dedalus_index: 0
-  test_batch_size: 2
-
