--- cfno/layers/normalization_layers.py
@@ -1,19 +1,0 @@
-import torch
-import torch.nn as nn
-
-class InstanceNorm(nn.Module):
-    def __init__(self, **kwargs):
-        """InstanceNorm applies dim-agnostic instance normalization
-        to data as an nn.Module. 
-
-        kwargs: additional parameters to pass to instance_norm() for use as a module
-        e.g. eps, affine
-        """
-        super().__init__()
-        self.kwargs = kwargs
-    
-    def forward(self, x):
-        size = x.shape
-        x = torch.nn.functional.instance_norm(x, **self.kwargs)
-        assert x.shape == size
-        return x
