+++ cfno/models/__init__.py
@@ -1,3 +1,1 @@
-from cfno.models.fno2d import FNO2D
-from cfno.models.fno3d import FNO3D
 from cfno.models.cfno2d import CFNO2D
