+++ cfno/training/__init__.py
@@ -1,2 +1,1 @@
-from cfno.training.origin import Trainer
 from cfno.training.pySDC import FourierNeuralOp
