--- docs/examples/requirements.txt
@@ -1,238 +1,0 @@
-absl-py==2.1.0
-alabaster==0.7.13
-appdirs==1.4.4
-arxiv_latex_cleaner==1.0.8
-asn1crypto==1.5.1
-asttokens==2.2.1
-atomicwrites==1.4.1
-attrs==23.1.0
-Babel==2.12.1
-backcall==0.2.0
-backports.entry-points-selectable==1.2.0
-backports.functools-lru-cache==1.6.5
-beautifulsoup4==4.12.2
-beniget==0.4.1
-bitstring==4.0.2
-bleach==6.1.0
-blist==1.3.6
-boto3==1.28.70
-botocore==1.31.70
-Bottleneck==1.3.7
-build==0.10.0
-CacheControl==0.12.14
-cachy==0.3.0
-certifi==2023.5.7
-cffi==1.15.1
-chardet==5.1.0
-charset-normalizer==3.1.0
-cleo==2.0.1
-click==8.1.3
-cloudpickle==2.2.1
-colorama==0.4.6
-commonmark==0.9.1
-configmypy==0.2.0
-contourpy==1.1.0
-crashtest==0.4.1
-cryptography==41.0.1
-cycler==0.11.0
-Cython==0.29.35
-deap==1.4.0
-decorator==5.1.1
-defusedxml==0.7.1
-dill==0.3.9
-distlib==0.3.6
-distro==1.8.0
-docopt==0.6.2
-docutils==0.20.1
-doit==0.36.0
-dulwich==0.21.5
-ecdsa==0.18.0
-editables==0.3
-exceptiongroup==1.1.1
-execnet==1.9.0
-executing==1.2.0
-expecttest==0.1.5
-fastjsonschema==2.19.1
-filelock==3.12.2
-flann==1.9.2
-flann==1.9.2
-flit_core==3.9.0
-fonttools==4.42.0
-fsspec==2023.6.0
-future==0.18.3
-gast==0.5.4
-glob2==0.7
-gmpy2==2.1.5
-grpcio==1.62.1
-h5py==3.9.0
-hatch-fancy-pypi-readme==23.1.0
-hatch-vcs==0.3.0
-hatchling==1.18.0
-html5lib==1.1
-idna==3.4
-imageio==2.31.3
-imagesize==1.4.1
-importlib-metadata==6.7.0
-importlib-resources==5.12.0
-iniconfig==2.0.0
-installer==0.7.0
-intervaltree==3.1.0
-intreehooks==1.0
-ipaddress==1.0.23
-ipython==8.12.3
-jaraco.classes==3.2.3
-jedi==0.19.0
-jeepney==0.8.0
-Jinja2==3.1.2
-jmespath==1.0.1
-joblib==1.2.0
-jsonschema==4.17.3
-jupyter_client==8.6.1
-jupyter_core==5.7.2
-jupyterlab_pygments==0.3.0
-keyring==23.13.1
-keyrings.alt==4.2.0
-kiwisolver==1.4.4
-liac-arff==2.5.0
-llvmlite==0.43.0
-lockfile==0.12.2
-lxml==4.9.2
-Markdown==3.6
-markdown-it-py==3.0.0
-MarkupSafe==2.1.3
-matplotlib==3.7.2
-matplotlib-inline==0.1.6
-mdurl==0.1.2
-meson==1.1.1
-meson-python==0.13.2
-mistune==3.0.2
-mock==5.0.2
-more-itertools==9.1.0
-mpi4py==3.1.4
-mpmath==1.3.0
-msgpack==1.0.5
-nbclient==0.10.0
-nbconvert==7.16.3
-nbformat==5.10.4
-netaddr==0.8.0
-netifaces==0.11.0
-networkx==3.1
-numba==0.60.0
-numexpr==2.8.4
-numpy==1.25.1
-packaging==23.1
-pandas==2.0.3
-pandocfilters==1.5.1
-parso==0.8.3
-pastel==0.2.1
-pathlib2==2.3.7.post1
-pathspec==0.11.1
-pbr==5.11.1
-pexpect==4.8.0
-pickleshare==0.7.5
-Pillow==9.5.0
-pip==23.1.2
-pipreqs==0.5.0
-pkginfo==1.9.6
-platformdirs==3.8.0
-pluggy==1.2.0
-ply==3.11
-poetry==1.5.1
-poetry-core==1.6.1
-poetry-plugin-export==1.4.0
-pooch==1.7.0
-prompt-toolkit==3.0.39
-protobuf==5.26.1
-psutil==5.9.5
-ptyprocess==0.7.0
-pure-eval==0.2.2
-py==1.11.0
-py-expression-eval==0.3.14
-pyasn1==0.5.0
-pybind11==2.11.1
-pycparser==2.21
-pycryptodome==3.18.0
-pydevtool==0.3.0
-Pygments==2.15.1
-pylev==1.4.0
-PyNaCl==1.5.0
-pyparsing==3.1.0
-pyproject_hooks==1.0.0
-pyproject-metadata==0.7.1
-pyrsistent==0.19.3
-pySDC==5.5.2
-pytest==7.4.0
-pytest-mock==3.14.0
-pytest-xdist==3.3.1
-python-dateutil==2.8.2
-pythran==0.13.1
-pytoml==0.1.21
-pytz==2023.3
-PyYAML==6.0
-pyzmq==25.1.2
-qmat==0.1.10
-rapidfuzz==2.15.1
-regex==2023.6.3
-requests==2.31.0
-requests-toolbelt==1.0.0
-rich==13.4.2
-rich-click==1.6.1
-ruamel.yaml==0.18.6
-ruamel.yaml.clib==0.2.8
-s3transfer==0.7.0
-scandir==1.10.0
-scipy==1.11.1
-seaborn==0.12.2
-SecretStorage==3.3.3
-semantic-version==2.10.0
-setuptools==67.8.0
-setuptools-scm==7.1.0
-shellingham==1.5.0
-simplegeneric==0.8.1
-simplejson==3.19.1
-six==1.16.0
-snowballstemmer==2.2.0
-sortedcontainers==2.4.0
-soupsieve==2.4.1
-Sphinx==7.0.1
-sphinx-bootstrap-theme==0.8.1
-sphinxcontrib-applehelp==1.0.4
-sphinxcontrib-devhelp==1.0.2
-sphinxcontrib-htmlhelp==2.0.1
-sphinxcontrib-jsmath==1.0.1
-sphinxcontrib-qthelp==1.0.3
-sphinxcontrib-serializinghtml==1.1.5
-sphinxcontrib-websupport==1.2.4
-stack-data==0.6.2
-sympy==1.12
-tabulate==0.9.0
-tensorboard==2.17.1
-tensorboard-data-server==0.7.2
-threadpoolctl==3.1.0
-tinycss2==1.2.1
-toml==0.10.2
-tomli==2.0.1
-tomli_w==1.0.0
-tomlkit==0.11.8
-torch==2.1.2
-torchsummary==1.5.1
-torch-dct
-tornado==6.4
-tqdm==4.66.1
-traitlets==5.9.0
-trove-classifiers==2023.5.24
-typing_extensions==4.6.3
-tzdata==2023.3
-ujson==5.8.0
-urllib3==1.26.16
-versioneer==0.29
-virtualenv==20.23.1
-wcwidth==0.2.6
-webencodings==0.5.1
-Werkzeug==3.0.2
-wheel==0.40.0
-xarray==2023.8.0
-xlrd==2.0.1
-yarg==0.1.9
-zipfile36==0.1.3
-zipp==3.15.0
