--- docs/examples/utils.py
@@ -1,128 +1,0 @@
-import os
-import math
-import torch
-import torch.nn as nn
-import numpy as np
-from math import sqrt
-import imageio
-
-units = {
-    0: 'B',
-    1: 'KiB',
-    2: 'MiB',
-    3: 'GiB',
-    4: 'TiB'
-}
-
-
-def format_mem(x):
-    """
-    Takes integer 'x' in bytes and returns a number in [0, 1024) and
-    the corresponding unit.
-
-    """
-    if abs(x) < 1024:
-        return round(x, 2), 'B'
-
-    scale = math.log2(abs(x)) // 10
-    scaled_x = x / (1024 ** scale)
-    unit = units[scale]
-
-    if int(scaled_x) == scaled_x:
-        return int(scaled_x), unit
-
-    # rounding leads to 2 or fewer decimal places, as required
-    return round(scaled_x, 2), unit
-
-def print_tensor_mem(x, id_str=None):
-    """
-    Prints the memory required by tensor 'x'.
-
-    """
-    if not CudaMemoryDebugger.ENABLED:
-        return
-
-    desc = 'memory'
-
-    if id_str is not None:
-        desc += f' ({id_str})'
-
-    desc += ':'
-
-    val, unit = format_mem(x.element_size() * x.nelement())
-
-    print(f'{desc} {val}{unit}')
-
-def format_tensor_size(x):
-    val, unit = format_mem(x)
-    return f'{val}{unit}'
-
-
-class CudaMemoryDebugger():
-    """
-    Helper to track changes in CUDA memory.
-
-    """
-    DEVICE = 'cuda'
-    LAST_MEM = 0
-    ENABLED = True
-
-
-    def __init__(self, print_mem):
-        self.print_mem = print_mem
-        if not CudaMemoryDebugger.ENABLED:
-            return
-
-        cur_mem = torch.cuda.memory_allocated(CudaMemoryDebugger.DEVICE)
-        cur_mem_fmt, cur_mem_unit = format_mem(cur_mem)
-        print(f'cuda allocated (initial): {cur_mem_fmt:.2f}{cur_mem_unit}')
-        CudaMemoryDebugger.LAST_MEM = cur_mem
-
-    def print(self,id_str=None):
-        if not CudaMemoryDebugger.ENABLED:
-            return
-
-        desc = 'cuda allocated'
-
-        if id_str is not None:
-            desc += f' ({id_str})'
-
-        desc += ':'
-
-        cur_mem = torch.cuda.memory_allocated(CudaMemoryDebugger.DEVICE)
-        cur_mem_fmt, cur_mem_unit = format_mem(cur_mem)
-
-        diff = cur_mem - CudaMemoryDebugger.LAST_MEM
-        if self.print_mem:
-            if diff == 0:
-                print(f'{desc} {cur_mem_fmt:.2f}{cur_mem_unit} (no change)')
-
-            else:
-                diff_fmt, diff_unit = format_mem(diff)
-                print(f'{desc} {cur_mem_fmt:.2f}{cur_mem_unit}'
-                      f' ({diff_fmt:+}{diff_unit})')
-
-        CudaMemoryDebugger.LAST_MEM = cur_mem
-
-
-def activation_selection(choice):
-    if choice in ['tanh', 'Tanh']:
-        return nn.Tanh()
-    elif choice in ['relu', 'ReLU']:
-        return nn.ReLU(inplace=True)
-    elif choice in ['lrelu', 'LReLU']:
-        return nn.LeakyReLU(inplace=True)
-    elif choice in ['sigmoid', 'Sigmoid']:
-        return nn.Sigmoid()
-    elif choice in ['softplus', 'Softplus']:
-        return nn.Softplus(beta=4)
-    elif choice in ['celu', 'CeLU']:
-        return nn.CELU()
-    elif choice in ['elu']:
-        return nn.ELU()
-    elif choice in ['mish']:
-        return nn.Mish()
-    else:
-        raise ValueError('Unknown activation function')
-    
-
